requirements.txt
openai~=1.12.0
python-dotenv~=1.0.1
gunicorn~=21.2.0
flask~=3.0.2